+++ src/windows/SQLite3-WinRT/SQLite3/Database.h
@@ -12,7 +12,7 @@ namespace SQLite3
     Database(Platform::String^ dbPath);
     virtual ~Database();
 
-    int close();
+    int closedb();
     int close_v2();
 
     Statement^ Prepare(Platform::String^ sql);